% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_RFCurveArray.R
\name{create_RFCurveArray}
\alias{create_RFCurveArray}
\title{Create Multidimensional Curve Arrays from File input}
\usage{
create_RFCurveArray(files)
}
\arguments{
\item{files}{\link{list} (\strong{required}): list of \code{.rf} files to be imported}
}
\value{
Returns a list with two arrays for the \code{RF_nat} and the \code{RF_reg} curve
}
\description{
Helper function to create a multidimensional curve array to
prepare the Bayesian process
}
\section{Function version}{
 0.1.0
}

\examples{

##TODO

}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2020. create_RFCurveArray(): Create Multidimensional Curve Arrays from File input. Function version 0.1.0. In: Kreutzer, S., Mittelstrass, D., 2020. RLumSTARR: Spatially Resolved Radiofluorescence Analysis. R package version 0.1.0.9000-53. 
}

