% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_TRUELight.R
\name{extract_TRUELight}
\alias{extract_TRUELight}
\title{Extract True Light from the Camera Measurements using a Bayesian
Approach}
\usage{
extract_TRUELight(
  data,
  element = c("RF_nat", "RF_reg"),
  ROI = 2,
  stepping = 1,
  method_control = list(),
  verbose = TRUE
)
}
\arguments{
\item{data}{\link{array} (\strong{required}): object created by \link{create_RFCurveArray}}

\item{element}{\link{character} (\emph{with default}): element from the input to be analysed,
'supported are only \code{RF_nat} or \code{RF_reg}}

\item{ROI}{\link{numeric} (\emph{optional}): ROI to be analysed, if nothing is given
all ROIs are analysed, however, the first ROIS is discarded!}

\item{stepping}{\link{numeric} (\emph{with default}): stepping parameter that allows
you to model only every xth (the value in \code{stepping}). This option can be extremely
useful to play with data because it dramatically improves the modelling speed because
less data are considered.}

\item{method_control}{\link{list} (\emph{optional}): parameter to be passed to \code{rjags}.
Supported are \code{n.chain}, \code{n.iter}, \code{thin}, \code{variable.names}, \code{model}, see details for more.}

\item{verbose}{\link{logical} (\emph{with default}): enable/disable terminal feedback}
}
\value{
Returns a list of class \code{RLumSTARR_TRUELight} with an the following elements:

\code{...$RF_curve}: \link[Luminescence:RLum.Data.Curve-class]{Luminescence::RLum.Data.Curve} object
(the RF curve with the true light)

\code{...$rjags_output}: \link[rjags:coda.samples]{rjags::coda.samples} output for further processing.
\emph{Note: Regardless the observed variable, the parameter alpha
will always be used to create the curve}

\code{...$model}: the model used to run the Bayesian process, use \link{writeLines} to have
nicely formatted terminal output
}
\description{
A Bayesian modelling approach to extract the true light using the
expanding region-of-interest (ROI) approach proposed by Cunningham and Clark-Balzan (2017).
The function will return the results for a \strong{single} curve from a \strong{single} ROI
}
\details{
\strong{Method control}

Supported options to be passed via the parameter \code{method_control}, most
of them are used internally for the calls to \link[runjags:run.jags]{runjags::run.jags} and \link[rjags:coda.samples]{rjags::coda.samples}.

\tabular{llll}{
\strong{PARAMETER} \tab \strong{TYPE} \tab \strong{DEFAULT} \tab \strong{DESCRIPTION}\cr
\code{n.chain} \tab \link{numeric} \tab \code{3} \tab the number of MCMC chains \cr
\code{thin} \tab \link{numeric} \tab \code{10} \tab thinning interval used for the monitoring\cr
\code{burnin} \tab \link{numeric} \tab \code{4000} \tab number of burn-in iterations, cf. \link[runjags:run.jags]{runjags::run.jags} \cr
\code{sample} \tab \link{numeric} \tab \code{10000} \tab number of total number of samples, \code{0} skips the MCMC sampling \cr
\code{adapt}  \tab \link{numeric} \tab  \code{1000} \tab number of adaptive iterations, cf. \link[runjags:run.jags]{runjags::run.jags} \cr
\code{summarise} \tab \link{logical} \tab \code{FALSE} \tab calculate output statistics \cr
\code{method} \tab \link{character} \tab \code{"rjags"} \tab way \code{runjags} will call JAGS, possible are \code{'rjags'}, \code{'simple'}, \code{'interruptible'}, \code{'parallel'}, \code{'rjparallel'}, \code{'background'}, \code{'bgparallel'} or \code{'snow'} cf. \link[runjags:run.jags]{runjags::run.jags} \cr
jags.refresh\verb{\\tab [numeric] \\tab}0.1\verb{\\tab refresh rate of update of the iteration process, select larger values of complex and long runs \\cr}variable.names\verb{\\tab [character] \\tab}"alpha"\verb{\\tab variable names to monitor,}alpha\verb{is always monitored \\cr}model` \tab \link{character} \tab the default model \tab the bugs model
}

\emph{Note: The argument \code{model} allows to heavily modified the underlying model. To avoid crashes the paramters passed by \code{variable.names} will always be cross-checked against parameteres present in the model. Unknown parameters will be skipped!}
}
\section{Function version}{
 0.1.1
}

\examples{

## list files using package external data
files <- list.files(system.file("extdata", "", package="RLumSTARR"), full.names=TRUE)

## create curve array
dat <- create_RFCurveArray(files = files)
output <- extract_TRUELight(
 data = dat,
 ROI = c(4),
 stepping = 10,
 verbose = FALSE,
 method_control = list(
   n.chain = 2,
   sample = 100,
   thin = 20))

} 

\section{How to cite}{
Kreutzer, S., 2022. extract_TRUELight(): Extract True Light from the Camera Measurements using a Bayesian Approach. Function version 0.1.1. In: Kreutzer, S., Mittelstrass, D., 2022. RLumSTARR: Spatially Resolved Radiofluorescence Analysis (EXPERIMENTAL PACKAGE). R package version 0.1.0.9000-132. 
}

\references{
Cunningham, A.C., Clark-Balzan, L., 2017. Overcoming crosstalk in
luminescence images of mineral grains. Radiation Measurements 106, 498–505.
doi:10.1016/j.radmeas.2017.06.004
}
\seealso{
\link{create_RFCurveArray}, \link{get_MCMCParameters}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
\keyword{datagen}
