% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MCMCParameters.R
\name{get_MCMCParameters}
\alias{get_MCMCParameters}
\title{Extracts a parameter from an MCMC list}
\usage{
get_MCMCParameters(mcmc, parameter = NULL, prob = 0.95, unlist = TRUE)
}
\arguments{
\item{mcmc}{\link[coda:mcmc]{coda::mcmc} or \link[coda:mcmc.list]{coda::mcmc.list} (\strong{required}): input object,
if created by \link{extract_TRUELight} the correct object is extracted automatically}

\item{parameter}{\link{character} (\emph{optional}): name of the parameter to be extracted. If \code{NULL} (the default)
all found parameters are extracted and the result is a list with matrices of those parameter. If \code{parameter}
is a vector, the function will try to extract the names parameters.}

\item{prob}{\link{numeric} (\emph{with default}): probability for the HPD calculation (cf. \link[coda:HPDinterval]{coda::HPDinterval})}

\item{unlist}{\link{logical} (\emph{with default}): if \code{TRUE} the output is a \link{matrix} of the means
of the lower and upper intervals of the parameter. If the parameter was estimated based on
multiple chains, this chains are also subject to an average calculation. If \code{FALSE} the output is a
\link{list} as returned by \link[coda:HPDinterval]{coda::HPDinterval}}
}
\value{
Returns a matrix with the parameter value or a named \link{list} with such matrices if \code{parameters} has
a length > 1.
}
\description{
Short cut to extract a parameter from an MCMC list or an object of class
\code{RLumSTARR_TRUELight} created by the functions \link{extract_TRUELight} or \link{run_TRUELightExtraction}.
In the latter case, the row names of the extracted matrices have the same dimension as the RF curve.
If more processing is wanted, \code{'coda'} package should be used.
}
\section{Function version}{
 0.1.0
}

\examples{

## load example files
files <- list.files(system.file("extdata", "", package="RLumSTARR"), full.names=TRUE)

##prepare data and run model
dat <- create_RFCurveArray(files = files)
output <-
 extract_TRUELight(
   data = dat,
   ROI = c(4),
   stepping = 60,
   verbose = FALSE,
   method_control = list(
   n.chain = 1,
   n.iter = 50,
   thin = 20))

##extract parameters
get_MCMCParameters(output)

}
\seealso{
\link[coda:HPDinterval]{coda::HPDinterval}, \link{extract_TRUELight}
}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2021. get_MCMCParameters(): Extracts a parameter from an MCMC list. Function version 0.1.0. In: Kreutzer, S., Mittelstrass, D., 2021. RLumSTARR: Spatially Resolved Radiofluorescence Analysis (EXPERIMENTAL PACKAGE). R package version 0.1.0.9000-108. 
}

