% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_RLumSTARR.R
\name{melt_RLumSTARR}
\alias{melt_RLumSTARR}
\title{Melt output from RLumSTARR into simple data frames}
\usage{
melt_RLumSTARR(x, ...)
}
\arguments{
\item{x}{(\strong{required}): input object of class \code{RLumSTARR_RFCurveArray} or
\code{RLumSTARR_TRUELight}}

\item{...}{further parameters, currently not used}
}
\value{
\link{data.frame}
}
\description{
The function provides a convenient way to convert the objects
created by RLumSTARR into simple \link{data.frame}s that can be processed conveniently by other
functions for example the package \code{ggplot2}.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University
(United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2022. melt_RLumSTARR(): Melt output from RLumSTARR into simple data frames. Function version 0.1.0. In: Kreutzer, S., Mittelstrass, D., 2022. RLumSTARR: Spatially Resolved Radiofluorescence Analysis (EXPERIMENTAL PACKAGE). R package version 0.1.0.9000-130. 
}

