% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methods_RLumSTARR}
\alias{methods_RLumSTARR}
\alias{plot.RLumSTARR_RFCurveArray}
\alias{print.RLumSTARR_model}
\title{methods_RLumSTARR}
\usage{
\method{plot}{RLumSTARR_RFCurveArray}(x, y = NULL, ...)

\method{print}{RLumSTARR_model}(x, ...)
}
\arguments{
\item{x:}{input object}

\item{y:}{input object}
}
\description{
Methods for S3-generics implemented for the package 'RLumSTARR'.
}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2022. methods_RLumSTARR(): methods_RLumSTARR. In: Kreutzer, S., Mittelstrass, D., 2022. RLumSTARR: Spatially Resolved Radiofluorescence Analysis (EXPERIMENTAL PACKAGE). R package version 0.1.0.9000-120. 
}

\keyword{internal}
