% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLumSTARR-package.R
\docType{package}
\name{RLumSTARR-package}
\alias{RLumSTARR-package}
\alias{RLumSTARR}
\title{SpaTiAlly Resolved Radiofluorescence}
\description{
A collection of functions to analyse spatially resolved radiofluorescence
data
}
\details{
\bold{Funding}
\itemize{
\item Sebastian Kreutzer received funding from the European Union’s Horizon 2020
research and innovation programme under the Marie Skłodowska-Curie grant agreement No 844457 (CREDit).
}
}
\references{
##TODO
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany) \cr
Dirk Mittelstrass (Germany)
, RLum Developer Team}
\keyword{package}
