% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ImageJ.R
\name{run_ImageJ}
\alias{run_ImageJ}
\title{Run ImageJ SR-RF macro}
\usage{
run_ImageJ(
  path,
  RF_nat = "default",
  RF_reg = "default",
  bg_rm = "take_from_RF_reg",
  image_group_size = 5,
  image_alignment = TRUE,
  first_slices_rm = FALSE,
  noise_tolorance = 10,
  ROI_size = 10,
  center_x = 0.5,
  center_y = 0.5,
  diameter = 0.9,
  use_predefined_ROIs = FALSE,
  channel_time = 5,
  save_workflow_images = FALSE,
  save_additional_results = FALSE,
  save_signal_decay_videos = FALSE,
  offset_time = 0,
  .ImageJ = "/Applications/Fiji.app/Contents/MacOS/ImageJ-macosx"
)
}
\arguments{
\item{path}{\link{character} (\strong{required}): path to files to be analysed}

\item{RF_nat}{\link{character} (\emph{default}): name of the RF_nat file}

\item{RF_reg}{\link{character} (\emph{default}): name of the RF_reg file}

\item{bg_rm}{\link{character} (\emph{with default}): background subtraction options.
Allowed are \code{none} (no background subtraction), \code{take_from_RF_reg} (takes the
last 100 channels from the RF_reg signal: dangerous) or \verb{<your file name>} (this
does not work in batch mode)}

\item{image_group_size}{\link{numeric} (\emph{with default}): grouping value for running median
to remove outliers}

\item{image_alignment}{\link{logical} (\emph{with default}): enable/disable image alignment}

\item{first_slices_rm}{\link{logical} (\emph{with default}): remove first slice of each curve set}

\item{noise_tolorance}{\link{numeric} (\emph{with default}): noise tolerance parameter}

\item{ROI_size}{\link{numeric} (\emph{with default}): ROI size in pixel}

\item{center_x}{\link{numeric} (\emph{with default}): aliquot ROI centre x-coordinate}

\item{center_y}{\link{numeric} (\emph{with default}): aliquot ROI centre y-coordinate}

\item{diameter}{\link{numeric} (\emph{with default}): relative diameter aliquot ROI}

\item{use_predefined_ROIs}{\link{logical} (\emph{with default}): use pre-defined ROIs imported
from a file \code{ROIs.zip} found in the same folder as the files}

\item{channel_time}{\link{numeric} (\emph{with default}): channel time, this parameter was set the
moment the sequence was written}

\item{save_workflow_images}{\link{logical} (\emph{with default}): enable/disable writing of additional workflow images}

\item{save_additional_results}{\link{logical} (\emph{with default}): enable/disable writing of additional workflow images}

\item{save_signal_decay_videos}{\link{logical} (\emph{with default}): enable/disable writing of additional workflow videos}

\item{offset_time}{\link{numeric} (\emph{with default}): offset time for the time axis}

\item{.ImageJ}{\link{numeric} (\emph{with default}): Path to ImageJ (the macro is shipped with the package)}
}
\value{
This functions returns the path of the analysed data
}
\description{
The script runs the SR-RF ImageJ macro in batch mode out of
R
}
\section{Function version}{
 0.1.0
}

\examples{

##TODO

}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2021. run_ImageJ(): Run ImageJ SR-RF macro. Function version 0.1.0. In: Kreutzer, S., Mittelstrass, D., 2021. RLumSTARR: Spatially Resolved Radiofluorescence Analysis (EXPERIMENTAL PACKAGE). R package version 0.1.0.9000-101. 
}

