% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TRUELightExtraction.R
\name{run_TRUELightExtraction}
\alias{run_TRUELightExtraction}
\title{Run Automated TRUE-Light Extraction over multiple datasets}
\usage{
run_TRUELightExtraction(
  data,
  ROI,
  stepping = 1,
  mc.cores = max(c(1, parallel::detectCores() - 2)),
  method_control = list(),
  include_jags_output = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{\link{array} (\strong{required}): object created by \link{create_RFCurveArray}}

\item{ROI}{\link{numeric} (\emph{optional}): ROIs to be analysed, if nothing is given
all ROIs are analysed, however, the first ROIS is discarded!}

\item{stepping}{\link{numeric} (\emph{with default}): the stepping parameter from and to
be passed to \link{extract_TRUELight}}

\item{mc.cores}{\link{numeric} (\emph{with default}): number of cores used for the processing,
passed to \link[parallel:mclapply]{parallel::mclapply}}

\item{method_control}{\link{list} (\emph{optional}): parameters to be passed to \link{extract_TRUELight}}

\item{include_jags_output}{\link{logical} (\emph{with default}): allows to include the output
from JAGS as info object in the \link[Luminescence:RLum.Analysis-class]{Luminescence::RLum.Analysis} output objects
for further diagnostics. Setting this option to \code{TRUE} is not recommended for large datasets,
since it will tremendously inflate the size of output and consume a lot of memory.
If in-depth diagnostics are required, the function \link{extract_TRUELight} is recommended
instead.}

\item{verbose}{\link{logical} (\emph{with default}): enable/disable verbose mode. The output of
the MCMC sampling using \code{rjags} is \strong{always} silent.}
}
\value{
The output is a \link{list} containing \link[Luminescence:RLum.Analysis-class]{Luminescence::RLum.Analysis} objects
with two \link[Luminescence:RLum.Data.Curve-class]{Luminescence::RLum.Data.Curve} objects for \code{RF_nat} and \code{RF_reg} respectively.
}
\description{
Runs \link{extract_TRUELight} automatically in a parallel processing
mode using \link[parallel:mclapply]{parallel::mclapply}. In essence, the function calls \link{extract_TRUELight}
and extracts the RF curves from the results and constructs an \link[Luminescence:RLum.Analysis-class]{Luminescence::RLum.Analysis}
for further analyses.
}
\section{Function version}{
 0.1.0
}

\examples{

## list files using package external data
files <- list.files(system.file("extdata", "", package="RLumSTARR"), full.names=TRUE)
## create curve array
dat <- create_RFCurveArray(files = files)
output <- run_TRUELightExtraction(
data = dat,
stepping = 15,
mc.cores = 1,
ROI = 5,
verbose = TRUE,
method_control = list(
 n.chain = 1,
 n.iter = 50,
 thin = 20))

}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2022. run_TRUELightExtraction(): Run Automated TRUE-Light Extraction over multiple datasets. Function version 0.1.0. In: Kreutzer, S., Mittelstrass, D., 2022. RLumSTARR: Spatially Resolved Radiofluorescence Analysis (EXPERIMENTAL PACKAGE). R package version 0.1.0.9000-118. 
}

