% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_TRUELight.R
\name{extract_TRUELight}
\alias{extract_TRUELight}
\title{Extract True Light from the Camera measurements using a Bayesian
Approach}
\usage{
extract_TRUELight(data, ROI = 1, method_control = list())
}
\arguments{
\item{data}{\link{array} (\strong{required}): curve array created by \link{create_RFCurveArray}}

\item{ROI}{\link{numeric} (\emph{with default}): ROI to be analysed}

\item{method_control}{\link{list} (\emph{optional}): parameter to be passed to \code{rjags}.
Supported are \code{n.chain}, \code{n.iter}, \code{thin}, \code{variable.names}, \code{model}}
}
\value{
Returns a list with an \link[Luminescence:RLum.Data.Curve-class]{Luminescence::RLum.Data.Curve} object
(the RF curve with the true light) and the \link[rjags:coda.samples]{rjags::coda.samples} output for
further processing. \emph{Note: Regardless the observed variable, the parameter alpha
will be always be used to create the curve}
}
\description{
Run the Bayesian model to extract the true light from a ROI
}
\section{Function version}{
 0.1.0
}

\examples{

##TODO

}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2020. extract_TRUELight(): Extract True Light from the Camera measurements using a Bayesian Approach. Function version 0.1.0. In: Kreutzer, S., Mittelstrass, D., 2020. RLumSTARR: Spatially Resolved Radiofluorescence Analysis. R package version 0.1.0.9000-29. 
}

