% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TRUELightExtraction.R
\name{run_TRUELightExtraction}
\alias{run_TRUELightExtraction}
\title{Run Automated TRUE-Light Extraction over multiple datasets}
\usage{
run_TRUELightExtraction(
  data,
  ROI,
  stepping = 1,
  mc.cores = max(c(1, parallel::detectCores() - 2)),
  method_control = list(),
  include_jags_output = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{[array] (**required**): object created by [create_RFCurveArray]}

\item{ROI}{[numeric] (*optional*): ROIs to be analysed, if nothing is given
all ROIs are analysed, however, the first ROIS is discarded!}

\item{stepping}{[numeric] (*with default*): the stepping parameter from and to
be passed to [extract_TRUELight]}

\item{mc.cores}{[numeric] (*with default*): number of cores used for the processing,
passed to [parallel::mclapply]}

\item{method_control}{[list] (*optional*): parameters to be passed to [extract_TRUELight]}

\item{include_jags_output}{[logical] (*with default*): allows to include the output
from JAGS as info object in the [Luminescence::RLum.Analysis-class] output objects
for further diagnostics. Setting this option to `TRUE` is not recommended for large datasets,
since it will tremendously inflate the size of output and consume a lot of memory.
If in-depth diagnostics are required, the function [extract_TRUELight] is recommended
instead.}

\item{verbose}{[logical] (*with default*): enable/disable verbose mode. The output of
the MCMC sampling using `rjags` is **always** silent.}
}
\value{
The output is a [list] containing [Luminescence::RLum.Analysis-class] objects
with two [Luminescence::RLum.Data.Curve-class] objects for `RF_nat` and `RF_reg` respectively.
}
\description{
Runs [extract_TRUELight] automatically in a parallel processing
mode using [parallel::mclapply]. In essence, the function calls [extract_TRUELight]
and extracts the RF curves from the results and constructs an [Luminescence::RLum.Analysis-class]
for further analyses.
}
\section{Function version}{
 0.1.0
}

\examples{

## list files using package external data
files <- list.files(system.file("extdata", "", package="RLumSTARR"), full.names=TRUE)
## create curve array
dat <- create_RFCurveArray(files = files)
output <- run_TRUELightExtraction(
data = dat,
stepping = 15,
mc.cores = 1,
ROI = 5,
verbose = TRUE,
method_control = list(
 n.chain = 1,
 n.iter = 50,
 thin = 20))

}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2021. run_TRUELightExtraction(): Run Automated TRUE-Light Extraction over multiple datasets. Function version 0.1.0. In: Kreutzer, S., Mittelstrass, D., 2021. RLumSTARR: Spatially Resolved Radiofluorescence Analysis (EXPERIMENTAL PACKAGE). R package version 0.1.0.9000-113. 
}

